Update README.md
Using for loops with an array

Step 1

Examine and run the code to see how it functions. Add a comment to the code each time the grades array is used, describing what that line of code is doing. You should have at least 4 comments added.

Step 2

Modify the code to process 4 array elements instead of 5.

Step 3

Add statements to also calculate the average of the 4 elements inputted and display the average to the screen.